Fix change in repo

+++ bin/test-wheels-macos.sh
@@ -6,35 +6,38 @@ set -e
 # Show explicitly which commands are currently running.
 set -x
 
-SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE:-$0}")"; pwd)
-WHEEL_DIR=$SCRIPT_DIR/../dist
+if [ -z $PYPI_REPOSITORY_URL ]
+then
+   SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE:-$0}")"; pwd)
+   WHEEL_DIR=$SCRIPT_DIR/../dist
 
-# Test whether we can install without any dependencies
-# TODO: delete once we have setup.py setup correctly
-for VERSION in $VERSIONS
-do
-  MAJOR=${VERSION%.*}
-  PYTHON_DIR=$(dirname $(realpath $(which python$MAJOR)))
-  PYTHON=$PYTHON_DIR/python$MAJOR
-  WHEEL_VERSION=$(echo $MAJOR | tr -d '.')
+   # Test whether we can install without any dependencies
+   # TODO: delete once we have setup.py setup correctly
+   for VERSION in $VERSIONS
+   do
+     MAJOR=${VERSION%.*}
+     PYTHON_DIR=$(dirname $(realpath $(which python$MAJOR)))
+     PYTHON=$PYTHON_DIR/python$MAJOR
+     WHEEL_VERSION=$(echo $MAJOR | tr -d '.')
 
-  git clean -f -f -x -d -q -e dist
+     git clean -f -f -x -d -q -e dist
 
-  $PYTHON -m venv venv
-  source venv/bin/activate
+     $PYTHON -m venv venv
+     source venv/bin/activate
 
-  # Find the appropriate wheel by grepping for the Python version.
-  MPI4PY_WHEEL=$(find $WHEEL_DIR -type f -maxdepth 1 -print | grep "$WHEEL_VERSION" | grep mpi4py)
+     # Find the appropriate wheel by grepping for the Python version.
+     MPI4PY_WHEEL=$(find $WHEEL_DIR -type f -maxdepth 1 -print | grep "$WHEEL_VERSION" | grep mpi4py)
 
-  # TODO: this will actually pick up both wheels since brainiak is in the path
-  BRAINIAK_WHEEL=$(find $WHEEL_DIR -type f -maxdepth 1 -print | grep "$WHEEL_VERSION" | grep brainiak)
+     # TODO: this will actually pick up both wheels since brainiak is in the path
+     BRAINIAK_WHEEL=$(find $WHEEL_DIR -type f -maxdepth 1 -print | grep "$WHEEL_VERSION" | grep brainiak)
 
-  $PYTHON -m pip install $MPI4PY_WHEEL
-  $PYTHON -m pip install $BRAINIAK_WHEEL
+     $PYTHON -m pip install $MPI4PY_WHEEL
+     $PYTHON -m pip install $BRAINIAK_WHEEL
 
-  deactivate
-  rm -rf venv
-done
+     deactivate
+     rm -rf venv
+   done
+fi
 
 brew install mpich
 
@@ -46,5 +49,11 @@ do
   PYTHON_DIR=$(dirname $(realpath $(which python$MAJOR)))
   PYTHON=$PYTHON_DIR/python$MAJOR
   git clean -f -f -x -d -q -e dist
-  WHEEL_DIR=$WHEEL_DIR PYTHON=$PYTHON $SCRIPT_DIR/pr-check.sh
+
+  if [ -z $PYPI_REPOSITORY_URL ]
+  then
+     WHEEL_DIR=$WHEEL_DIR PYTHON=$PYTHON $SCRIPT_DIR/pr-check.sh
+  else
+     PYTHON=$PYTHON $SCRIPT_DIR/pr-check.sh
+  fi
 done
